install.sh

+++ install.sh
@@ -208,17 +208,32 @@ then
                			fi
                 	fi
 		#Creating an icon in the desktop
-
-		if [[ -f "$home/Desktop/Linux_Burner.desktop" ]];
+		if [[ -d "$home/Desktop" ]];
 		then
-			rm $home/Desktop/Linux_Burner.desktop 
-			printf "[Desktop Entry]\nVersion=1.0\nType=Application\nName=LinuxBurner\nComment=System monitoring\nExec=bash -c \"cd %s/python && python3 interface.py\"\nIcon=$REP/assets/icones/monitoring.png\nTerminal=true" "$REP"  > $home/Linux.Burner.desktop
-			echo -e "${Yellow}Giving the right permission for the app ${Reset}"
-			sudo -S chmod +x $home/Linux_Burner.desktop <<< $PASS
-		else
-			printf "[Desktop Entry]\nVersion=1.0\nType=Application\nName=LinuxBurner\nComment=System monitoring\nExec=bash -c \"cd %s/python && python3 interface.py\"\nIcon=$REP/assets/icones/monitoring.png\nTerminal=true" "$REP"  > $home/Desktop/Linux_Burner.desktop
-			echo -e "${Yellow}Giving the right permission for the app${Reset}"
-			sudo -S chmod +x $home/Linux_Burner.desktop <<< $PASS
+			if [[ -f "$home/Desktop/Linux_Burner.desktop" ]];
+			then
+				rm $home/Desktop/Linux_Burner.desktop 
+				printf "[Desktop Entry]\nVersion=1.0\nType=Application\nName=LinuxBurner\nComment=System monitoring\nExec=bash -c \"cd %s/python && python3 interface.py\"\nIcon=$REP/assets/icones/monitoring.png\nTerminal=true" "$REP"  > $home/Desktop/Linux.Burner.desktop
+				echo -e "${Yellow}Giving the right permission for the app ${Reset}"
+				sudo -S chmod +x $home/Desktop/Linux_Burner.desktop <<< $PASS
+			else
+				printf "[Desktop Entry]\nVersion=1.0\nType=Application\nName=LinuxBurner\nComment=System monitoring\nExec=bash -c \"cd %s/python && python3 interface.py\"\nIcon=$REP/assets/icones/monitoring.png\nTerminal=true" "$REP"  > $home/Desktop/Linux_Burner.desktop
+				echo -e "${Yellow}Giving the right permission for the app${Reset}"
+				sudo -S chmod +x $home/Desktop/Linux_Burner.desktop <<< $PASS
+			fi
+		elif [[ -d "$home/Bureau" ]];
+		then
+			if [[ -f "$home/Bureau/Linux_Burner.desktop" ]];
+                        then
+                                rm $home/Desktop/Linux_Burner.desktop
+                                printf "[Desktop Entry]\nVersion=1.0\nType=Application\nName=LinuxBurner\nComment=System monitoring\nExec=bash -c \"cd %s/python && python3 interface.py\"\nIcon=$REP/assets/icones/monitoring.png\nTerminal=true" "$REP"  > $home/Bureau/Linux.Burner.desktop
+                                echo -e "${Yellow}Giving the right permission for the app ${Reset}"
+                                sudo -S chmod +x $home/Bureau/Linux_Burner.desktop <<< $PASS
+                        else
+                                printf "[Desktop Entry]\nVersion=1.0\nType=Application\nName=LinuxBurner\nComment=System monitoring\nExec=bash -c \"cd %s/python && python3 interface.py\"\nIcon=$REP/assets/icones/monitoring.png\nTerminal=true" "$REP"  > $home/Bureau/Linux_Burner.desktop
+                                echo -e "${Yellow}Giving the right permission for the app${Reset}"
+                                sudo -S chmod +x $home/Bureau/Linux_Burner.desktop <<< $PASS
+                        fi
 		fi
 
 		echo -e "${Red}
